move date range checker (#3132)
