+++ services/ui-src/src/routes/Admin/Admin.tsx
@@ -9,7 +9,7 @@ import {
   ErrorAlert,
 } from "components";
 // utils
-import { checkBannerActivityStatus, formatDateUtcToEt } from "utils";
+import { checkDateRangeStatus, formatDateUtcToEt } from "utils";
 import { DELETE_BANNER_FAILED } from "verbiage/errors";
 import verbiage from "verbiage/admin-view";
 
@@ -17,7 +17,7 @@ export const Admin = () => {
   const { bannerData, deleteAdminBanner, writeAdminBanner, errorMessage } =
     useContext(AdminBannerContext);
   const [error, setError] = useState<string | undefined>(errorMessage);
-  const bannerIsActive = checkBannerActivityStatus(
+  const bannerIsActive = checkDateRangeStatus(
     bannerData?.startDate,
     bannerData?.endDate
   );
