--- services/ui-src/src/utils/banners/adminBanner.ts
@@ -1,7 +1,0 @@
-export const checkBannerActivityStatus = (
-  startDate: number,
-  endDate: number
-): boolean => {
-  const currentTime = new Date().valueOf();
-  return currentTime >= startDate && currentTime <= endDate;
-};
