+++ services/ui-src/src/utils/other/time.ts
@@ -80,3 +80,11 @@ export const convertDatetimeStringToNumber = (
   }
   return convertedTime || undefined;
 };
+
+export const checkDateRangeStatus = (
+  startDate: number,
+  endDate: number
+): boolean => {
+  const currentTime = new Date().valueOf();
+  return currentTime >= startDate && currentTime <= endDate;
+};
